Refactor test structure and add gRPC support with new test scripts and data files
