+++ package.json
@@ -10,8 +10,9 @@
     "build": "webpack",
     "lint": "eslint src && prettier --check src",
     "lint-fix": "eslint src --fix && prettier --write src",
-    "test": "node src/test.js",
-    "k6-simple": "k6 run tests/simple.js"
+    "test": "node tests/harness.js",
+    "k6-simple": "k6 run tests/k6/simple.js",
+    "k6-grpc": "k6 run tests/k6/grpc.js"
   },
   "devDependencies": {
     "@eslint/js": "^9.37.0",
