+++ tests/k6/grpc.js
@@ -3,17 +3,17 @@ import { check } from 'k6'
 import { Counter, Trend, Rate } from 'k6/metrics'
 
 //import { htmlReport } from "https://raw.githubusercontent.com/benc-uk/k6-reporter/main/dist/bundle.js";
-import { htmlReport } from '../dist/bundle.js'
+import { htmlReport } from '../../dist/bundle.js'
 import { textSummary } from 'https://jslib.k6.io/k6-summary/0.1.0/index.js'
 
 const client = new grpc.Client()
 client.load(['.'], 'hello.proto')
 
-var grpcReqCounter = new Counter('grpc_reqs')
-var fakeTrend = new Trend('fake_trend')
-var fakeRate = new Rate('fake_rate')
+const grpcReqCounter = new Counter('grpc_reqs')
+const fakeTrend = new Trend('fake_trend')
+const fakeRate = new Rate('fake_rate')
 
-export let options = {
+export const options = {
   stages: [{ duration: '2s', target: 5 }],
   thresholds: {
     grpc_req_duration: ['med <= 2'],
