+++ tests/k6/simple.js
@@ -6,7 +6,7 @@ import { check, group, sleep } from 'k6'
 import http from 'k6/http'
 
 //import { htmlReport } from "https://raw.githubusercontent.com/benc-uk/k6-reporter/main/dist/bundle.js";
-import { htmlReport } from '../dist/bundle.js'
+import { htmlReport } from '../../dist/bundle.js'
 import { textSummary } from 'https://jslib.k6.io/k6-summary/0.1.0/index.js'
 
 const TARGET_URL = __ENV.TEST_TARGET || 'https://benc.dev'
